requirements.txt
streamlit
streamlit-webrtc
ultralytics
opencv-python-headless
Pillow
torch
